@@ -71,11 +74,11 @@ THE SOFTWARE.
         </pluginRepository>
     </pluginRepositories>
     <properties>
-        <checkstyle.version>2.9.1</checkstyle.version>
-        <cobertura.version>2.5.2</cobertura.version>
+        <jenkins.version>1.655</jenkins.version>
+        <workflow.version>1.11</workflow.version>
         <powermock.version>1.5.5</powermock.version>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
         <maven-hpi-plugin.version>1.110</maven-hpi-plugin.version>
         <npm-executable>npm</npm-executable>
     </properties>
@@ -121,9 +142,8 @@ THE SOFTWARE.
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
-                    <debug>true</debug>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
             <plugin>
