@@ -24,26 +24,29 @@ THE SOFTWARE.
 package com.groupon.jenkins.github;
 
 import com.groupon.jenkins.dynamic.build.DynamicProject;
-import com.groupon.jenkins.dynamic.build.cause.GithubCause;
+import com.groupon.jenkins.dynamic.build.cause.GithubPushPullWebhookCause;
 import com.groupon.jenkins.dynamic.build.repository.DynamicProjectRepository;
 import hudson.model.ParameterDefinition;
 import hudson.model.ParametersAction;
 import hudson.model.ParametersDefinitionProperty;
 import hudson.model.StringParameterValue;
 import hudson.security.ACL;
-import java.io.IOException;
-import java.util.Arrays;
 import org.acegisecurity.context.SecurityContextHolder;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.kohsuke.github.GHEvent;
 import org.kohsuke.stapler.StaplerRequest;
 import org.mockito.ArgumentCaptor;
 import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 
+import java.io.IOException;
+import java.util.Arrays;
+
 import static com.google.common.collect.Lists.newArrayList;
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyString;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
@@ -62,85 +65,88 @@ public class GithubWebhookTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void should_not_be_called_with_payload() throws IOException {
-        githubWebhook.doIndex(mock(StaplerRequest.class), null);
+        this.githubWebhook.doIndex(mock(StaplerRequest.class), null);
     }
 
     @Test
     public void should_get_payload_from_post_if_post() throws IOException, InterruptedException {
-        StaplerRequest request = mock(StaplerRequest.class);
-        when(request.getParameter("payload")).thenReturn("payload");
+        final StaplerRequest request = mock(StaplerRequest.class);
+        when(request.getParameter("payload")).thenReturn("{repository: { url: 'git@repo' },deleted: false}");
+        when(request.getHeader("X-GitHub-Event")).thenReturn(GHEvent.PUSH.toString());
         when(request.getMethod()).thenReturn("POST");
-        DynamicProject project = mock(DynamicProject.class);
-        ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
+        final DynamicProject project = mock(DynamicProject.class);
+        final ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
         when(project.getProperty(ParametersDefinitionProperty.class)).thenReturn(paramDefinition);
 
         kickOffBuildTrigger(request, project);
 
-        verify(project).scheduleBuild(eq(0), any(GithubCause.class), any(ParametersAction.class));
+        verify(project).scheduleBuild(eq(0), any(GithubPushPullWebhookCause.class), any(ParametersAction.class));
     }
 
     @Test
     public void should_trigger_builds_for_payload() throws IOException, InterruptedException {
-        StaplerRequest request = mock(StaplerRequest.class);
+        final StaplerRequest request = mock(StaplerRequest.class);
         when(request.getParameter("payload")).thenReturn("payload");
-        DynamicProject project = mock(DynamicProject.class);
-        ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
+        final DynamicProject project = mock(DynamicProject.class);
+        final ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
         when(project.getProperty(ParametersDefinitionProperty.class)).thenReturn(paramDefinition);
 
         kickOffBuildTrigger(request, project);
 
-        verify(project).scheduleBuild(eq(0), any(GithubCause.class), any(ParametersAction.class));
+        verify(project).scheduleBuild(eq(0), any(GithubPushPullWebhookCause.class), any(ParametersAction.class));
     }
 
     @Test
     public void should_trigger_build_with_default_parameter_values() throws IOException, InterruptedException {
-        StaplerRequest request = mock(StaplerRequest.class);
+        final StaplerRequest request = mock(StaplerRequest.class);
         when(request.getParameter("payload")).thenReturn("payload");
-        DynamicProject project = mock(DynamicProject.class);
+        final DynamicProject project = mock(DynamicProject.class);
 
-        ParameterDefinition branchParameter = mock(ParameterDefinition.class);
+        final ParameterDefinition branchParameter = mock(ParameterDefinition.class);
         when(branchParameter.getName()).thenReturn("BRANCH");
 
-        ParameterDefinition secondParameter = mock(ParameterDefinition.class);
+        final ParameterDefinition secondParameter = mock(ParameterDefinition.class);
         when(branchParameter.getName()).thenReturn("PARAM");
-        when(secondParameter.getDefaultParameterValue()).thenReturn(new StringParameterValue("PARAM","meow"));
-        ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
-        when(paramDefinition.getParameterDefinitions()).thenReturn(Arrays.asList(branchParameter,secondParameter));
+        when(secondParameter.getDefaultParameterValue()).thenReturn(new StringParameterValue("PARAM", "meow"));
+        final ParametersDefinitionProperty paramDefinition = mock(ParametersDefinitionProperty.class);
+        when(paramDefinition.getParameterDefinitions()).thenReturn(Arrays.asList(branchParameter, secondParameter));
         when(project.getProperty(ParametersDefinitionProperty.class)).thenReturn(paramDefinition);
 
         kickOffBuildTrigger(request, project);
 
-        ArgumentCaptor<ParametersAction> parametersCaptor = ArgumentCaptor.forClass(ParametersAction.class);
-        verify(project).scheduleBuild(eq(0), any(GithubCause.class), parametersCaptor.capture());
+        final ArgumentCaptor<ParametersAction> parametersCaptor = ArgumentCaptor.forClass(ParametersAction.class);
+        verify(project).scheduleBuild(eq(0), any(GithubPushPullWebhookCause.class), parametersCaptor.capture());
 
-        ParametersAction parametersAction = parametersCaptor.getValue();
-        Assert.assertTrue(parametersAction.getParameter("PARAM") instanceof  StringParameterValue);
+        final ParametersAction parametersAction = parametersCaptor.getValue();
+        Assert.assertTrue(parametersAction.getParameter("PARAM") instanceof StringParameterValue);
         Assert.assertEquals("meow", ((StringParameterValue) parametersAction.getParameter("PARAM")).value);
 
     }
+
     @Test
     public void should_authenticate_as_SYSTEM() throws IOException, InterruptedException {
-        StaplerRequest request = mock(StaplerRequest.class);
-        DynamicProject project = mock(DynamicProject.class);
+        final StaplerRequest request = mock(StaplerRequest.class);
+        final DynamicProject project = mock(DynamicProject.class);
         when(request.getParameter("payload")).thenReturn("payload");
-        kickOffBuildTrigger(request,project);
+        kickOffBuildTrigger(request, project);
         Assert.assertEquals(ACL.SYSTEM, SecurityContextHolder.getContext().getAuthentication());
     }
 
 
-    protected void kickOffBuildTrigger(StaplerRequest request, DynamicProject projectForRepo) throws IOException, InterruptedException {
-        Payload payload = mock(Payload.class);
+    protected void kickOffBuildTrigger(final StaplerRequest request, final DynamicProject projectForRepo) throws IOException, InterruptedException {
+        final PushAndPullRequestPayload payload = mock(PushAndPullRequestPayload.class);
 
-        DynamicProjectRepository projectRepo = mock(DynamicProjectRepository.class);
+        final DynamicProjectRepository projectRepo = mock(DynamicProjectRepository.class);
 
-        when(payload.needsBuild(false)).thenReturn(true);
+        when(payload.needsBuild(any(DynamicProject.class))).thenReturn(true);
         when(payload.getProjectUrl()).thenReturn("git@repo");
+        when(payload.getCause()).thenReturn(mock(GithubPushPullWebhookCause.class));
         when(projectRepo.getJobsFor("git@repo")).thenReturn(newArrayList(projectForRepo));
 
-        doReturn(payload).when(githubWebhook).makePayload("payload");
-        doReturn(projectRepo).when(githubWebhook).makeDynamicProjectRepo();
+        doReturn(payload).when(this.githubWebhook).makePayload(anyString(), anyString());
+        doReturn(projectRepo).when(this.githubWebhook).makeDynamicProjectRepo();
 
-        githubWebhook.doIndex(request, null);
+        this.githubWebhook.doIndex(request, null);
         Thread.sleep(2000);
     }
 
